requirements.txt
python-binance
pandas
python-dotenv
ccxt
sqlalchemy
psycopg2-binary